:sparkles: bug(all): fixes bugs
