+++ client/scripts/endGame.js
@@ -1,4 +1,5 @@
 import closePopup from './lib/closePopup.mjs';
+import displayNewGame from './displayNewGame.js';
 
 export default function endGame(
   socket,
@@ -62,37 +63,7 @@ export default function endGame(
   popupBackground.appendChild(popup);
   document.body.appendChild(popupBackground);
 
-  const gridContainer = document.querySelector('.grid_container');
-  gridContainer.remove();
-  timerContainer.remove();
-
-  const waitingSpan = document.createElement('span');
-  waitingSpan.innerText = 'Waiting for 2-4 players to connect...';
-  waitingSpan.classList.add('waiting_span');
-
-  beforeGameContainer.appendChild(waitingSpan);
-
-  //Listen to if theres 2-4 players in room
-  socket.emit('start over', room);
-  socket.on('start over', () => {
-    beforeGameContainer.innerHTML = '';
-    const startGameBtn = document.createElement('button');
-    startGameBtn.innerText = 'Play again';
-
-    startGameBtn.addEventListener('click', () => {
-      socket.emit('startCountdown', room);
-    });
-    beforeGameContainer.append(startGameBtn);
-    //If someone leaves
-    socket.on('player left', () => {
-      beforeGameContainer.innerHTML = '';
-      const waitingSpan = document.createElement('span');
-      waitingSpan.innerText = 'Waiting for 2-4 players to connect...';
-      waitingSpan.classList.add('waiting_span');
-
-      beforeGameContainer.appendChild(waitingSpan);
-    });
-  });
+  displayNewGame(socket, room, beforeGameContainer, timerContainer);
 
   gameContainer.appendChild(beforeGameContainer);
 }
