+++ server/lib/gameRoom.js
@@ -49,10 +49,28 @@ const gameRoom = {
       }
     });
 
+    function calculateNull(gameGrid) {
+      const allCellsNull = gameGrid.every((row) =>
+        row.every((cell) => cell === null)
+      );
+
+      if (allCellsNull) {
+        return true;
+      } else {
+        return false;
+      }
+    }
+
     //Listen to when time is up.
     socket.on('endGame', async (room) => {
-      if (gameStarted) {
-        await endGameSession(room.roomId);
+      if (calculateNull(gameGrids[room.roomId])) {
+        console.log('NO CELL CLICKED!');
+        gameStarted = false;
+        io.to(room.roomId).emit('game without click');
+      } else {
+        if (gameStarted) {
+          await endGameSession(room.roomId);
+        }
       }
     });
 
@@ -95,7 +113,7 @@ const gameRoom = {
     });
 
     let gameStarted = false;
-    let gameDuration = 1 * 45 * 1000;
+    let gameDuration = 1 * 5 * 1000;
 
     async function endGameSession(roomId) {
       gameStarted = false;
